[ignore] Script/Changelog: provide write method
